Chore : 컴포넌트 분리
두 개의 컴포넌트를 서로 분리하였습니다.
